equation: merged to parsing, setup: added universal flag
