+++ docal/__init__.py
@@ -1,6 +1,6 @@
 '''initializiation module'''
 
-from .equation import eqn
+from .parsing import eqn
 from .calculation import cal
 from .formatting import format_quantity as fmt
 from .document import document
